review: factor out attribute conversion into FromHoudini class
